+++ ChangeLog.md
@@ -3,6 +3,10 @@ XP Reflection ChangeLog
 
 ## ?.?.? / ????-??-??
 
+* Fixed `Type` class to consistently raise CannotInstantiate exceptions
+  for interfaces and abstract classes
+  (@thekid)
+
 ## 0.1.0 / 2020-12-07
 
 * Hello World! First release - @thekid